# swagger.yaml
{
"swagger": "2.0",
"info": {
"description": "Api Documentation",
"version": "1.0",
"title": "Api Documentation",
"termsOfService": "urn:tos",
"contact": {},
"license": {
"name": "Apache 2.0",
"url": "http://www.apache.org/licenses/LICENSE-2.0"
}
},
"host": "localhost:8080",
"basePath": "/",
...
"paths": {
...
"/invite": {
"get": {
"tags": [
"invite-controller"
],
"summary": "getInviteByGoal",
"operationId": "getInviteByGoalUsingGET",
"consumes": [
"application/json"
],
"produces": [
"*/*"
],
"parameters": [
{
"name": "goal",
"in": "query",
"description": "goal",
"required": true,
"type": "string"
}
],
"responses": {
"200": {
"description": "OK",
"schema": {
"$ref": "#/definitions/Invite"
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
},
"404": {
"description": "Not Found"
}
}
},
"post": {
"tags": [
"invite-controller"
],
"summary": "createInvite",
"operationId": "createInviteUsingPOST",
"consumes": [
"application/json"
],
"produces": [
"*/*"
],
"parameters": [
{
"in": "body",
"name": "invite",
"description": "invite",
"required": true,
"schema": {
"$ref": "#/definitions/Invite"
}
}
],
"responses": {
"201": {
"description": "Created",
"schema": {
"$ref": "#/definitions/Invite"
}
},
"401": {
"description": "Unauthorized"
},
"403": {
"description": "Forbidden"
},
"404": {
"description": "Not Found"
}
}
}
},
...
},
"definitions": {
...
"Invite": {
"type": "object",
"properties": {
"goal": {
"type": "string"
},
"guid": {
"type": "string"
}
}
}
}
}